requirements.txt
fastapi
uvicorn
huggingface_hub
llama-cpp-python